Refactor to use powerpack style db for page ingestion
This cleans up how pages and content are defined and loaded in a way
that requires less code to add new pages.

Notably I did not use powerpack (a static site tool by cjohansen)
because it only generates static sites. Whereas this site should be
memory cached (so basically static), but will support some
interactive/live elements in the future.